Create config.py
The code seeks to extract the relevant configs from the config_path file path to be used in the main code. There is also a get_config function that seeks user input if the required input is not found in the config file or alternatively exits the code if specified. This would happen when user input is not easily obtained e.g. if the input is a list or a dictionary