Create main.py
This is the main file where the script is executed. It loads the config from the config file based on the filepath specified in the environment and runs the crawler based on the params provided in the config file. The sample config file can be found in this repo called pg_config.yaml. 

The output of the crawler is a list of properties that fit the criteria specified in the params in the config file. It will then check against the output csv specified. If there is no such file a new output csv will be created and the properties populated there. If there is an output csv file with data already existing, this will help the script identify which properties are new or which have been observed before in previous crawls

As this is a live crawl, please respect the robots.txt and do not abuse the script.